add repository, repository.common, service, service.common, model, model.common, common and dal folders

+++ Example.WebApi/Controllers/PokemonController.cs
@@ -1,6 +1,7 @@
 ﻿using Example.WebApi.Models;
 using Microsoft.Ajax.Utilities;
 using Npgsql;
+using NpgsqlTypes;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -12,6 +13,7 @@ using System.Net.Http;
 using System.Security.Cryptography;
 using System.Web;
 using System.Web.Http;
+using System.Web.UI;
 
 namespace Example.WebApi.Controllers
 {
@@ -34,14 +36,96 @@ namespace Example.WebApi.Controllers
         // GET: api/Pokemon
         public HttpResponseMessage Get([FromUri] PokemonRead pokemon)
         {
-            List<Pokemon> pokemons = new List<Pokemon>();
+            List<PokemonRead> pokemons = new List<PokemonRead>();
             NpgsqlConnection connection = new NpgsqlConnection(connectionString);
-            using (connection)
+
+            if (pokemon == null)
             {
-                NpgsqlCommand command = new NpgsqlCommand();
-                command.Connection = connection;
-                command.CommandText = $"SELECT \"Id\" , \"TrainerId\" , \"Name\", \"Type\" , \"SecondType\" FROM \"Pokemon\"";
-            }     
+                using (connection)
+                {
+                    connection.Open();
+                    NpgsqlCommand command = new NpgsqlCommand();
+                    command.Connection = connection;
+                    command.CommandText = $"SELECT * FROM \"Pokemon\" LEFT JOIN \"Trainer\" ON \"Trainer\".\"Id\" = \"TrainerId\"";
+
+                    using (NpgsqlDataReader reader = command.ExecuteReader())
+                    {
+                        while (reader.Read())
+                        {
+                            PokemonRead pokemonRead = new PokemonRead
+                            {
+                                PokemonId = reader["Id"] != DBNull.Value ? Convert.ToInt32(reader["Id"]) : 0,
+                                TrainerId = reader["TrainerId"] != DBNull.Value ? Convert.ToInt32(reader["TrainerId"]) : 0,
+                                Name = reader["Name"] != DBNull.Value ? reader["Name"].ToString() : string.Empty,
+                                Type = reader["Type"] != DBNull.Value ? reader["Type"].ToString() : string.Empty,
+                                SecondType = reader["SecondType"] != DBNull.Value ? reader["SecondType"].ToString() : string.Empty,
+                            };
+
+                            pokemons.Add(pokemonRead);
+                        }
+                    }
+                }
+            }
+            else
+            {
+                using (connection)
+                {
+                    connection.Open();
+                    NpgsqlCommand command = new NpgsqlCommand();
+                    command.Connection = connection;
+                    List<string> conditions = new List<string>();
+                    if (pokemon.Name != null)
+                    {
+                        string filter1 = $"\"Pokemon\".\"Name\" = @name";
+                        command.Parameters.AddWithValue("name", pokemon.Name);
+                        conditions.Add(filter1);
+                    }
+                    if (pokemon.Type != null)
+                    {
+                        string filter2 = $"\"Pokemon\".\"Type\" = @type";
+                        command.Parameters.AddWithValue("type", pokemon.Type);
+                        conditions.Add(filter2);
+                    }
+                    if (pokemon.SecondType != null)
+                    {
+                        string filter3 = $"\"Pokemon\".\"SecondType\" = @secondType";
+                        command.Parameters.AddWithValue("secondType", pokemon.SecondType);
+                        conditions.Add(filter3);
+                    }
+                    if (pokemon.TrainerId != 0)
+                    {
+                        string filter4 = $"\"Pokemon\".\"TrainerId\" = @trainerId";
+                        command.Parameters.AddWithValue("trainerId", pokemon.TrainerId);
+                        conditions.Add(filter4);
+                    }
+                    string baseQuery = $"SELECT * FROM \"Pokemon\" LEFT JOIN \"Trainer\" ON \"Trainer\".\"Id\" = \"TrainerId\" WHERE 1=1 ";
+                    if (conditions.Count > 0)
+                    {
+                        baseQuery = baseQuery + "AND ";
+                    }
+                    string conditionsText = String.Join(" AND ", conditions);
+                    command.CommandText = baseQuery + conditionsText;
+                    using (NpgsqlDataReader reader = command.ExecuteReader())
+                    {
+                        while (reader.Read())
+                        {
+                            PokemonRead pokemonRead = new PokemonRead
+                            {
+                                PokemonId = reader["Id"] != DBNull.Value ? Convert.ToInt32(reader["Id"]) : 0,
+                                TrainerId = reader["TrainerId"] != DBNull.Value ? Convert.ToInt32(reader["TrainerId"]) : 0,
+                                Name = reader["Name"] != DBNull.Value ? reader["Name"].ToString() : string.Empty,
+                                Type = reader["Type"] != DBNull.Value ? reader["Type"].ToString() : string.Empty,
+                                SecondType = reader["SecondType"] != DBNull.Value ? reader["SecondType"].ToString() : string.Empty,
+                            };
+
+                            pokemons.Add(pokemonRead);
+                        }
+                    }
+
+
+                }
+            }
+            return Request.CreateResponse(HttpStatusCode.OK, pokemons);
         }
 
         [HttpGet]
@@ -59,9 +143,9 @@ namespace Example.WebApi.Controllers
                 command.Parameters.AddWithValue("id", id);
                 connection.Open();
                 NpgsqlDataReader reader = command.ExecuteReader();
-                reader.Read();
-                if(reader.HasRows)
+                if (reader.HasRows)
                 {
+                    reader.Read();
                     pokemon = new Pokemon(
                     (int)reader["Id"],
                     (int)reader["TrainerId"],
@@ -72,10 +156,74 @@ namespace Example.WebApi.Controllers
                 }
                 connection.Close();
             }
-            if(pokemon == null) {
+            if (pokemon == null)
+            {
                 return Request.CreateResponse(HttpStatusCode.NotFound, "Pokemon not found");
             }
             return Request.CreateResponse(HttpStatusCode.OK, pokemon);
+        }
+
+
+        private bool PokemonExists(int id)
+        {
+            Pokemon pokemon = null;
+            NpgsqlConnection connection = new NpgsqlConnection(connectionString);
+
+            using (connection)
+            {
+                NpgsqlCommand command = new NpgsqlCommand();
+                command.CommandText = $"SELECT * FROM \"Pokemon\" WHERE \"Id\" = @id";
+                command.Connection = connection;
+                command.Parameters.AddWithValue("id", id);
+                connection.Open();
+                NpgsqlDataReader reader = command.ExecuteReader();
+                if (reader.HasRows)
+                {
+                    reader.Read();
+                    pokemon = new Pokemon(
+                    (int)reader["Id"],
+                    (int)reader["TrainerId"],
+                    (string)reader["Name"],
+                    (string)reader["Type"],
+                    (string)reader["SecondType"]
+                );
+                }
+                connection.Close();
+            }
+            if (pokemon == null)
+            {
+                return false;
+            }
+            return true;
+        }
+
+        private Pokemon FetchPokemon(int id)
+        {
+            Pokemon pokemon = null;
+            NpgsqlConnection connection = new NpgsqlConnection(connectionString);
+
+            using (connection)
+            {
+                NpgsqlCommand command = new NpgsqlCommand();
+                command.CommandText = $"SELECT * FROM \"Pokemon\" WHERE \"Id\" = @id";
+                command.Connection = connection;
+                command.Parameters.AddWithValue("id", id);
+                connection.Open();
+                NpgsqlDataReader reader = command.ExecuteReader();
+                if (reader.HasRows)
+                {
+                    reader.Read();
+                    pokemon = new Pokemon(
+                    (int)reader["Id"],
+                    (int)reader["TrainerId"],
+                    (string)reader["Name"],
+                    (string)reader["Type"],
+                    (string)reader["SecondType"]
+                );
+                }
+                connection.Close();
+            }
+            return pokemon;
         }
 
 
@@ -115,26 +263,74 @@ namespace Example.WebApi.Controllers
         // PUT: api/Pokemon/5
         public HttpResponseMessage Put(int id, [FromBody] PokemonUpdate updatedPokemon)
         {
-            NpgsqlCommand command = new NpgsqlCommand();
-            command.CommandText = $"UPDATE \"Pokemon\" SET \"TrainerId\"=@trainerId, \"Name\"=@name, \"Type\"=@type, \"SecondType\"=@secondType WHERE \"Id\"=@id";
-            NpgsqlConnection connection = new NpgsqlConnection(connectionString);
-            command.Parameters.AddWithValue("id", id);
-            command.Parameters.AddWithValue("trainerId", updatedPokemon.TrainerId);
-            command.Parameters.AddWithValue("name", updatedPokemon.Name);
-            command.Parameters.AddWithValue("type", updatedPokemon.Type);
-            command.Parameters.AddWithValue("secondType", updatedPokemon.SecondType);
-            int numberOfAffectedRows;
-            using (connection)
+            if (updatedPokemon == null)
             {
-                connection.Open();
-                numberOfAffectedRows = command.ExecuteNonQuery();
-                connection.Close();
+                return Request.CreateResponse(HttpStatusCode.BadRequest);
             }
-            if (numberOfAffectedRows > 0)
+            else if (PokemonExists(id) == false)
             {
-                return Request.CreateResponse(HttpStatusCode.OK, "Pokemon updated");
+                return Request.CreateResponse(HttpStatusCode.NotFound, "Pokemon not found");
             }
-            return Request.CreateResponse(HttpStatusCode.BadRequest);
+            else
+            {
+
+                Pokemon pokemonToUpdate = FetchPokemon(id);
+                List<string> conditions = new List<string>();
+                NpgsqlCommand command = new NpgsqlCommand();
+                string baseQuery = $"UPDATE \"Pokemon\" SET ";
+                if (updatedPokemon.TrainerId != pokemonToUpdate.TrainerId)
+                {
+                    string condition1 = $"\"TrainerId\"=@trainerId";
+                    command.Parameters.AddWithValue("trainerId", updatedPokemon.TrainerId);
+                    conditions.Add(condition1);
+                }
+                if (updatedPokemon.Name != pokemonToUpdate.Name)
+                {
+                    string condition2 = $"\"Name\"=@name";
+                    command.Parameters.AddWithValue("name", updatedPokemon.Name);
+                    conditions.Add(condition2);
+                }
+                if (updatedPokemon.Type != pokemonToUpdate.Type)
+                {
+                    string condition3 = "\"Type\"=@type";
+                    command.Parameters.AddWithValue("type", updatedPokemon.Type);
+                    conditions.Add(condition3);
+                }
+                if (updatedPokemon.SecondType != pokemonToUpdate.SecondType)
+                {
+                    string condition4 = "\"SecondType\"=@secondType";
+                    command.Parameters.AddWithValue("secondType", updatedPokemon.SecondType);
+                    conditions.Add(condition4);
+                }
+                string whereClause = " WHERE \"Id\"=@id";
+                command.Parameters.AddWithValue("id", id);
+                command.CommandText = "";
+
+
+
+                if (conditions.Count == 0)
+                {
+                    return Request.CreateResponse(HttpStatusCode.BadRequest, "No changes made");
+                }
+                string conditionsText = String.Join(" , ", conditions);
+                command.CommandText = baseQuery + conditionsText + whereClause;
+
+
+                NpgsqlConnection connection = new NpgsqlConnection(connectionString);
+                command.Connection = connection;
+                int numberOfAffectedRows;
+                using (connection)
+                {
+                    connection.Open();
+                    numberOfAffectedRows = command.ExecuteNonQuery();
+                    connection.Close();
+                }
+                if (numberOfAffectedRows > 0)
+                {
+                    return Request.CreateResponse(HttpStatusCode.OK, "Pokemon updated");
+                }
+                return Request.CreateResponse(HttpStatusCode.BadRequest);
+            }
         }
 
 
@@ -142,25 +338,35 @@ namespace Example.WebApi.Controllers
         // DELETE: api/Pokemon/5
         public HttpResponseMessage Delete(int id)
         {
-            NpgsqlCommand command = new NpgsqlCommand();
-            NpgsqlConnection connection = new NpgsqlConnection(connectionString);
-            command.CommandText = $"DELETE FROM \"Pokemon\" WHERE \"Id\"=(@id)";
-            command.Parameters.AddWithValue("id", id);
-            int numberOfAffectedRows;
-            using (connection)
+            if (id == 0)
             {
-                connection.Open();
-                numberOfAffectedRows = command.ExecuteNonQuery();
-                connection.Close();
+                return Request.CreateResponse(HttpStatusCode.BadRequest);
             }
-
-            if (numberOfAffectedRows > 0)
+            else if (PokemonExists(id) == false)
             {
-                return Request.CreateResponse(HttpStatusCode.OK, "Pokemon deleted");
+                return Request.CreateResponse(HttpStatusCode.NotFound, "Pokemon not found");
             }
-            return Request.CreateResponse(HttpStatusCode.BadRequest, "Pokemon not deleted");
+            else
+            {
+                NpgsqlCommand command = new NpgsqlCommand();
+                NpgsqlConnection connection = new NpgsqlConnection(connectionString);
+                command.CommandText = $"DELETE FROM \"Pokemon\" WHERE \"Id\"=(@id)";
+                command.Parameters.AddWithValue("id", id);
+                command.Connection = connection;
+                int numberOfAffectedRows;
+                using (connection)
+                {
+                    connection.Open();
+                    numberOfAffectedRows = command.ExecuteNonQuery();
+                    connection.Close();
+                }
 
+                if (numberOfAffectedRows > 0)
+                {
+                    return Request.CreateResponse(HttpStatusCode.OK, "Pokemon deleted");
+                }
+                return Request.CreateResponse(HttpStatusCode.BadRequest, "Pokemon not deleted");
+            }
         }
-
     }
 }
